+++ src/main/java/vazkii/psi/common/spell/selector/PieceSelectorSavedVector.java
@@ -43,7 +43,7 @@ public class PieceSelectorSavedVector extends PieceSelector {
 		super.addToMetadata(meta);
 
 		Double numberVal = this.<Double>getParamEvaluation(number);
-		if(numberVal == null || numberVal <= 0 || numberVal.doubleValue() != numberVal.intValue())
+		if(numberVal == null || numberVal <= 0 || numberVal != numberVal.intValue())
 			throw new SpellCompilationException(SpellCompilationException.NON_POSITIVE_INTEGER, x, y);
 
 		meta.addStat(EnumSpellStat.POTENCY, numberVal.intValue() * 6);
