+++ src/main/java/vazkii/psi/common/spell/trick/entity/PieceTrickIgnite.java
@@ -42,7 +42,7 @@ public class PieceTrickIgnite extends PieceTrick {
 		super.addToMetadata(meta);
 		Double timeVal = this.<Double>getParamEvaluation(time);
 
-		if(timeVal == null ||  timeVal <= 0 || timeVal.doubleValue() != timeVal.intValue())
+		if(timeVal == null ||  timeVal <= 0 || timeVal != timeVal.intValue())
 			throw new SpellCompilationException(SpellCompilationException.NON_POSITIVE_INTEGER, x, y);
 
 		meta.addStat(EnumSpellStat.POTENCY, timeVal.intValue() * 40);
